Conflict chest and loot profile
Create Conflict chest, it is tied to the generator switch for spawning on death of governer. I will need to push new switches and landblock files later.

Create Loot Profile, based off Legendary chests and upped the quality of items, the chances the items will be of better quality etc.